HackTheArt 01.10.2026/CTF/llm/requirements.txt
flask
python-dotenv
google-generativeai>=0.3.0
